+++ spec/lib/config_spec.rb
@@ -529,9 +529,18 @@ RSpec.describe Doorkeeper::Config do
     it "can change the value" do
       Doorkeeper.configure do
         orm DOORKEEPER_ORM
-        custom_access_token_attributes [:added_field_1, :added_field_2]
+        custom_access_token_attributes [:tenant_name]
       end
-      expect(config.custom_access_token_attributes).to eq([:added_field_1, :added_field_2])
+      expect(config.custom_access_token_attributes).to eq([:tenant_name])
+    end
+
+    it "validates custom attributes to be present in the models" do
+      expect do
+        Doorkeeper.configure do
+          orm DOORKEEPER_ORM
+          custom_access_token_attributes [:none_existing_column]
+        end
+      end.to raise_error(Doorkeeper::Errors::ConfigError)
     end
   end