+++ CHANGELOG.md
@@ -8,6 +8,8 @@ User-visible changes worth mentioning.
 ## main
 
 - [#ID] Add your PR description here.
+- [#1602] Allow custom data to be stored inside access grants/tokens.
+- [#1603] Code refactoring for custom token attributes.
 
 # 5.6.4